Backend/requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cloudinary==1.22.0
cycler==0.10.0
Flask==1.1.2
Flask-Cors==3.0.9
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.2
noise==1.2.2
numpy==1.19.2
opencv-python-headless==4.4.0.44
Pillow==8.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.24.0
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
